requirements.txt
pandas
numpy
scikit-learn
yellowbrick
seaborn
pymysql
mysql_connector_python
python-dotenv
dataclasses